be/src/main/java/jshop/domain/user/controller/UserController.java
package jshop.domain.user.controller;

import jakarta.validation.Valid;
import jshop.domain.user.dto.JoinDto;
import jshop.domain.user.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;


@RestController
@RequiredArgsConstructor
public class UserController {

  private final UserService userService;

  @PostMapping("/api/join")
  @ResponseStatus(HttpStatus.NO_CONTENT)
  public void join(@RequestBody @Valid JoinDto joinDto) {
    userService.joinUser(joinDto);
  }
}

